requirements.txt
Django==1.9
MySQL-python==1.2.5
Pillow==3.3.0
PyYAML==3.12
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.2
appnope==0.1.0
argparse==1.4.0
billiard==3.3.0.23
celery==3.1.23
decorator==4.0.10
diff-match-patch==20121119
dj-database-url==0.4.1
dj-static==0.0.6
django-classy-tags==0.7.2
django-filter==0.14.0
django-import-export==0.5.0
django-mptt==0.8.5
django-reversion==2.0.6
django-sekizai==0.9.0
django-toolbelt==0.0.1
enum34==1.1.6
gnureadline==6.3.3
gunicorn==19.6.0
ipython==4.1.2
ipython-genutils==0.1.0
kombu==3.0.35
packaging==16.8
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
psycopg2==2.6.2
ptyprocess==0.5.1
py-mysql2pgsql==0.1.5
pyparsing==2.1.10
pytz==2016.6.1
redis==2.10.5
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.1
static3==0.7.0
tablib==0.11.2
termcolor==1.1.0
traitlets==4.3.1
unicode-slugify==0.1.1
whitenoise==3.2
wsgiref==0.1.2
